move tags into shared
