+++ src/app/shared/tags/tags-store.module.ts
@@ -2,13 +2,13 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
-import { TagsEffects } from './effects';
-import { reducer } from './reducer';
+import { TagsEffects } from './tags.effects';
+import { tagsReducer } from './tags.reducer';
 
 @NgModule({
   imports: [
     CommonModule,
-    StoreModule.forFeature('tags', reducer),
+    StoreModule.forFeature('tags', tagsReducer),
     EffectsModule.forFeature([TagsEffects]),
   ],
   providers: [TagsEffects],
