+++ tsconfig.json
@@ -13,9 +13,8 @@
     "target": "es2015",
     "module": "es2020",
     "paths": {
-      "@services/*": ["src/app/services/*"],
+      "@app/*": ["src/app/*"],
       "@store/*": ["src/app/app-store/*"],
-      "@models/*": ["src/app/models/*"],
       "@shared/*": ["src/app/shared/*"]
     },
     "lib": ["es2018", "dom"]
